g EM-517: Don't complain if asked to delete a file that doesn't exist
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@23128 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ modules/gateway/processcontroller/src/main/java/com/l7tech/server/processcontroller/ConfigServiceImpl.java
@@ -457,6 +457,11 @@ public class ConfigServiceImpl implements ConfigService {
         if (monitoringConfigFile == null) return;
 
         if (config == null) {
+            if (!monitoringConfigFile.exists()) {
+                logger.log(Level.INFO, "Monitoring configuration has been unset, but configuration file was already deleted");
+                return;
+            }
+
             try {
                 if (!FileUtils.deleteFileSafely(monitoringConfigFile.getPath())) {
                     logger.log(Level.WARNING, "Unable to delete unset monitoring configuration");
